Add files via upload
This project represents my first contribution to Github. This was done mainly to practice using binary classification models.

I was curious to see which classification model performed best on the dataset with regards to predicting customer satisfaction/dissatisfaction and the models performed as follows:

1) Random Forest             (95.7% accuracy)
2) Kernel SVM                   (95.4% accuracy)
3) Decision Tree                (94.9% accuracy)
4) XGBoost                       (94.0% accuracy)
5) KNN                             (92.8% accuracy)
6) SVM                             (87.4% accuracy)
7) Logistic Regression      (87.2% accuracy)
8) Naive Bayes                 (84.5% accuracy)

I was also sure to apply K-fold Cross Validation for each of the models to ensure consistency in the results.
Interestingly enough, the XGBoost model finished middle-of-the-pack, rather than surpassing the other models as expected.